+++ spec/support/mongoid.rb
@@ -25,7 +25,7 @@ class MongoidGrid
   end
 
   filter :name
-  integer_range_filters(:group_id, {:default => 0}, {:default => 100})
+  filter(:group_id, :integer, range: true, default: 0..100)
   filter :disabled, :xboolean
 
   column :name